+++ ninja-core/src/main/java/ninja/standalone/StandaloneHelper.java
@@ -17,13 +17,27 @@
 package ninja.standalone;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.ServerSocket;
+import java.net.URI;
+import java.net.URL;
+import java.security.KeyStore;
+import javax.net.ssl.KeyManager;
+import javax.net.ssl.KeyManagerFactory;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.TrustManager;
+import javax.net.ssl.TrustManagerFactory;
 import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Helper utilities for working with standalone applications.
  */
 public class StandaloneHelper {
+    static private final Logger log = LoggerFactory.getLogger(StandaloneHelper.class);
+    
+    static private final String URI_SCHEME_CLASSPATH = "classpath";
     
     static public int findAvailablePort(int min, int max) {
         for (int port = min; port < max; port++) {
@@ -77,6 +91,61 @@ public class StandaloneHelper {
             throw new IllegalArgumentException("A context path must not end with a '/' character. " +
                     " See https://docs.oracle.com/javaee/6/api/javax/servlet/ServletContext.html#getContextPath() for more info");
         }
+    }
+    
+    static public InputStream openKeyStoreInput(URI uri) throws IOException {
+        if (uri.getScheme().equals(URI_SCHEME_CLASSPATH)) {
+            String resourceName = uri.getPath();
+            
+            log.debug("Opening keystore on classpath with resource {}", resourceName);
+            
+            InputStream stream = StandaloneHelper.class.getResourceAsStream(resourceName);
+        
+            if (stream == null) {
+                throw new IOException("Resource '" + resourceName + "' not found on classpath");
+            }
+
+            return stream;
+        } else {
+            URL url = uri.toURL();
+            
+            log.debug("Opening keystore with url {}", url);
+            
+            return url.openStream();
+        }
+    }
+    
+    static public KeyStore loadKeyStore(URI uri, char[] password) throws Exception {
+        try (InputStream stream = openKeyStoreInput(uri)) {
+            KeyStore ks = KeyStore.getInstance("JKS");
+            ks.load(stream, password);
+            return ks;
+        }
+    }
+    
+    static public SSLContext createSSLContext(
+            URI keystoreUri, char[] keystorePassword,
+            URI truststoreUri, char[] truststorePassword) throws Exception {
+        
+        // load keystore
+        KeyStore keystore = loadKeyStore(keystoreUri, keystorePassword);
+        KeyManager[] keyManagers;
+        KeyManagerFactory keyManagerFactory = KeyManagerFactory.getInstance(KeyManagerFactory.getDefaultAlgorithm());
+        keyManagerFactory.init(keystore, keystorePassword);
+        keyManagers = keyManagerFactory.getKeyManagers();
+
+        // load truststore
+        KeyStore truststore = loadKeyStore(truststoreUri, truststorePassword);
+        TrustManager[] trustManagers;
+        TrustManagerFactory trustManagerFactory = TrustManagerFactory.getInstance(TrustManagerFactory.getDefaultAlgorithm());
+        trustManagerFactory.init(truststore);
+        trustManagers = trustManagerFactory.getTrustManagers();
+
+        SSLContext sslContext;
+        sslContext = SSLContext.getInstance("TLS");
+        sslContext.init(keyManagers, trustManagers, null);
+
+        return sslContext;
     }
     
 }