+++ ninja-core/src/main/java/ninja/utils/OverlayedNinjaProperties.java
@@ -17,6 +17,8 @@
 package ninja.utils;
 
 import com.google.inject.Singleton;
+import java.net.URI;
+import java.net.URISyntaxException;
 import javax.inject.Inject;
 import org.apache.commons.lang.StringUtils;
 
@@ -101,6 +103,21 @@ public class OverlayedNinjaProperties {
             return Long.valueOf(value);
         } catch (Exception e) {
             throw new IllegalArgumentException("Unable to convert property '" + key + "' with value '" + value + "' to a Long", e);
+        }
+    }
+    
+    public URI getURI(String key, URI currentValue, URI defaultValue) {
+        String value = null;
+        try {
+            value = get(key, safeToString(currentValue), safeToString(defaultValue));
+            
+            if (StringUtils.isEmpty(value)) {
+                return null;
+            }
+            
+            return new URI(value);
+        } catch (Exception e) {
+            throw new IllegalArgumentException("Unable to convert property '" + key + "' with value '" + value + "' to a URI", e);
         }
     }